Exception handling - delete

+++ src/main/java/com/hebertesteves/webservices/services/UserService.java
@@ -2,8 +2,11 @@ package com.hebertesteves.webservices.services;
 
 import com.hebertesteves.webservices.entities.User;
 import com.hebertesteves.webservices.repositories.UserRepository;
+import com.hebertesteves.webservices.services.exceptions.DatabaseException;
 import com.hebertesteves.webservices.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +32,16 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
